Extract ColumnWithFrostedGlassBackground as reusable component
